transfer logic from actor to non actor
