+++ Assets/Scripts/SS3D/Systems/Substances/SubstanceContainer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2385b8893215d534baf45728f30628dd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
