+++ Assets/Scripts/SS3D/Systems/Substances/SubstanceContainerActor.cs
@@ -17,66 +17,52 @@ namespace SS3D.Substances
     public class SubstanceContainerActor : InteractionTargetNetworkBehaviour
     {
 
-        /// <summary>
-        /// A list of all substances in this container
-        /// </summary>
-        public List<SubstanceEntry> Substances
-        {
-            get => substances;
-            private set => substances = value;
-        }
-
-        /// <summary>
-        /// The total number of moles
-        /// </summary>
-        public float TotalMoles => Substances.Sum(x => x.Moles);
-
-        /// <summary>
-        /// The filled volume in ml
-        /// </summary>
-        public float CurrentVolume => currentVolume;
 
         [SyncVar]
         private float currentVolume;
 
-        /// <summary>
-        /// The remaining volume in milliliters that fit in this container
-        /// </summary>
-        public float RemainingVolume => volume - currentVolume;
 
         [SyncVar]
         private float volume;
 
-        /// <summary>
-        /// Multiplier to convert moles in this container to volume
-        /// </summary>
-        public float MolesToVolume
-        {
-            get
-            {
-                float val = 0;
-                float total = TotalMoles;
-                foreach (var entry in Substances)
-                {
-                    val += entry.Substance.MillilitersPerMole * (entry.Moles / total);
-                }
-                return val;
-            }
-        }
+        private SubstanceContainer substanceContainer;
 
         /// <summary>
-        /// The capacity of this container in milliliters
+        /// A list of all substances in this container
         /// </summary>
-        public float Volume 
-        {
-            get => volume;
-        }
-        
+        [SerializeField]
+        public List<SubstanceEntry> InitialSubstances;
+
+        /// <summary>
+        /// The remaining volume in milliliters that fit in this container
+        /// </summary>
+        public float RemainingVolume => substanceContainer.RemainingVolume;
+
+        /// <summary>
+        /// The filled volume in ml
+        /// </summary>
+        public float CurrentVolume => substanceContainer.CurrentVolume;
 
         /// <summary>
         /// The temperature of the container
         /// </summary>
-        public float Temperature;
+        public float Temperature => substanceContainer.Temperature;
+
+        public float TotalMoles => substanceContainer.TotalMoles;
+
+        public List<SubstanceEntry> Substances => substanceContainer.Substances;
+
+
+        /// <summary>
+        /// The capacity of this container in milliliters
+        /// </summary>
+        public float Volume
+        {
+            get => substanceContainer.Volume;
+        }
+
+
+
 
         /// <summary>
         /// Is the container locked?
@@ -88,14 +74,9 @@ namespace SS3D.Substances
 
         public event OnContentChanged ContentsChanged;
 
-        [SerializeField] private List<SubstanceEntry> substances;
-
         private void Start()
         {
-            if (Substances.Count < 1)
-            {
-                Substances = new List<SubstanceEntry>();
-            }
+            
             if (IsServer)
             {
                 RecalculateAndSyncVolume();
@@ -113,16 +94,6 @@ namespace SS3D.Substances
         }
 
         /// <summary>
-        /// Can this container hold an additional amount of milliliters
-        /// </summary>
-        /// <param name="milliliters">The amount of additional milliliters</param>
-        /// <returns>If it fits the container</returns>
-        public bool CanFitUnits(uint milliliters)
-        {
-            return RemainingVolume >= milliliters;
-        }
-
-        /// <summary>
         /// Adds an amount of an substance to the container.
         /// </summary>
         /// <param name="substance">The substance to add</param>
@@ -130,30 +101,7 @@ namespace SS3D.Substances
         [Server]
         public void AddSubstance(Substance substance, float moles)
         {
-           
-            if (!CanTransfer())
-                return;
-            
-            var remainingCapacity = RemainingVolume;
-            var additionalVolume = moles * substance.MillilitersPerMole;
-            if (additionalVolume > remainingCapacity)
-            {
-                moles = remainingCapacity / substance.MillilitersPerMole;
-            }
-
-            int index = Substances.FindIndex(x => x.Substance == substance);
-            if (index == -1)
-            {
-                Substances.Add(new SubstanceEntry(substance, moles));
-            }
-            else
-            {
-                SubstanceEntry entry = Substances[index];
-                entry.Moles += moles;
-                Substances[index] = entry;
-            }
-            Debug.Log("substance is added. Remaining volume is " + RemainingVolume);
-
+            substanceContainer.AddSubstance(substance, moles);
             RecalculateAndSyncVolume();
         }
 
@@ -165,7 +113,7 @@ namespace SS3D.Substances
         [Server]
         public bool ContainsSubstance(Substance substance, float moles = 0.0001f)
         {
-            return Substances.FirstOrDefault(x => x.Substance == substance).Moles >= moles;
+            return substanceContainer.ContainsSubstance(substance, moles);
         }
 
         /// <summary>
@@ -176,26 +124,7 @@ namespace SS3D.Substances
         [Server]
         public void RemoveSubstance(Substance substance, float moles = float.MaxValue)
         {
-            if (!CanTransfer()) 
-                return;
-            
-            int index = IndexOfSubstance(substance);
-            if (index < 0)
-            {
-                return;
-            }
-
-            SubstanceEntry entry = Substances[index];
-            float newAmount = entry.Moles - moles;
-            if (newAmount <= 0.000001)
-            {
-                Substances.RemoveAt(index);
-            }
-            else
-            {
-                entry.Moles = newAmount;
-                Substances[index] = entry;
-            }
+            substanceContainer.RemoveSubstance(substance, moles);
             RecalculateAndSyncVolume();
         }
 
@@ -204,7 +133,7 @@ namespace SS3D.Substances
         /// </summary>
         public void Empty()
         {
-            Substances.Clear();
+            substanceContainer.Empty();
         }
 
         /// <summary>
@@ -214,31 +143,7 @@ namespace SS3D.Substances
         [Server]
         public void RemoveMoles(float moles)
         {
-            var totalMoles = Substances.Sum(x => x.Moles);
-            if (moles > totalMoles)
-            {
-                moles = totalMoles;
-            }
-
-            if (moles <= 0)
-            {
-                return;
-            }
-
-            for (var i = 0; i < Substances.Count; i++)
-            {
-                SubstanceEntry entry = Substances[i];
-                entry.Moles -= entry.Moles / totalMoles * moles;
-                if (entry.Moles <= 0.0001)
-                {
-                    Substances.RemoveAt(i);
-                    i--;
-                }
-                else
-                {
-                    Substances[i] = entry;
-                }
-            }
+            substanceContainer.RemoveMoles(moles);
         }
 
         /// <summary>
@@ -249,31 +154,7 @@ namespace SS3D.Substances
         [Server]
         public void TransferMoles(SubstanceContainerActor other, float moles)
         {
-            var totalMoles = Substances.Sum(x => x.Moles);
-            if (moles > totalMoles)
-            {
-                moles = totalMoles;
-            }
-
-            float relativeMoles = moles / totalMoles;
-
-            for (var i = 0; i < Substances.Count; i++)
-            {
-                SubstanceEntry entry = Substances[i];
-                float entryMoles = entry.Moles * relativeMoles;
-                entry.Moles -= entryMoles;
-                other.AddSubstance(entry.Substance, entryMoles);
-                if (entry.Moles <= 0.0000001)
-                {
-                    Substances.RemoveAt(i);
-                    i--;
-                }
-                else
-                {
-                    Substances[i] = entry;
-                }
-            }
-
+            substanceContainer.TransferMoles(other, moles);
             RecalculateAndSyncVolume();
         }
 
@@ -285,7 +166,7 @@ namespace SS3D.Substances
         [Server]
         public void TransferVolume(SubstanceContainerActor other, float milliliters)
         {
-            TransferMoles(other, milliliters / MolesToVolume);
+            substanceContainer.TransferVolume(other,milliliters);
         }
 
         /// <summary>
@@ -296,15 +177,7 @@ namespace SS3D.Substances
         [Server]
         public int IndexOfSubstance(Substance substance)
         {
-            for (int i = 0; i < Substances.Count; i++)
-            {
-                if (Substances[i].Substance == substance)
-                {
-                    return i;
-                }
-            }
-
-            return -1;
+            return substanceContainer.IndexOfSubstance(substance);
         }
 
         /// <summary>
@@ -323,7 +196,7 @@ namespace SS3D.Substances
         [Server]
         private void RecalculateAndSyncVolume()
         {
-            currentVolume = Substances.Sum(x => x.Moles * x.Substance.MillilitersPerMole);
+            currentVolume = substanceContainer.Substances.Sum(x => x.Moles * x.Substance.MillilitersPerMole);
         }
 
         [Server]
@@ -346,108 +219,8 @@ namespace SS3D.Substances
         /// </summary>
         public void ProcessContainer(SubstanceContainerActor container)
         {
-            var registry = Subsystems.Get<SubstancesSystem>();
-            float temperature = container.Temperature;
-
-            // Process recipes
-            // TODO : Highly inefficient as most recipes won't be achievable given the substances in the substance container.
-            // Instead, find a good data structure able to search through a small set of recipes, based on number of 
-            // ingredients or even ingredient types.
-            foreach (Recipe recipe in registry.Recipes)
-            {
-                // Check temperature limits
-                if (temperature < recipe.MinimalTemperature || temperature > recipe.MaximalTemperature)
-                {
-                    continue;
-                }
-
-                // Gather the mole amount of every substance
-                float[] moles = new float[recipe.Ingredients.Length];
-                bool ingredientsPresent = true;
-                for (var i = 0; i < recipe.Ingredients.Length; i++)
-                {
-                    ingredientsPresent = false;
-                    foreach (var entry in container.Substances)
-                    {
-                        if (entry.Substance.Type == recipe.Ingredients[i].Type)
-                        {
-                            moles[i] = entry.Moles;
-                            ingredientsPresent = true;
-                            break;
-                        }
-                    }
-
-                    if (!ingredientsPresent)
-                    {
-                        break;
-                    }
-                }
-
-                // Substance missing
-                if (!ingredientsPresent)
-                {
-                    continue;
-                }
-
-                // Calculate the maximum amount of ingredients
-                float totalIngredients = recipe.Ingredients.Sum(x => x.RelativeAmount);
-                float maxConversion = float.MaxValue;
-                for (var i = 0; i < moles.Length; i++)
-                {
-                    float relativeAmount = recipe.Ingredients[i].RelativeAmount;
-                    float part = relativeAmount / totalIngredients;
-                    float maxProduced = moles[i] / part;
-                    if (maxProduced < maxConversion)
-                    {
-                        maxConversion = maxProduced;
-                    }
-                }
-
-                // Calculate relative volume of ingredients
-                Substance[] ingredientSubstances = new Substance[moles.Length];
-                float ingredientsToVolume = 0;
-                for (var i = 0; i < moles.Length; i++)
-                {
-                    var component = recipe.Ingredients[i];
-                    var substance = ingredientSubstances[i] = registry.FromType(component.Type);
-                    ingredientsToVolume += substance.MillilitersPerMole * component.RelativeAmount;
-                }
-                ingredientsToVolume /= totalIngredients;
-
-                // Calculate relative volume of results
-                float totalResults = recipe.Results.Sum(x => x.RelativeAmount);
-                Substance[] resultSubstances = new Substance[recipe.Results.Length];
-                float resultsToVolume = 0;
-                for (var i = 0; i < recipe.Results.Length; i++)
-                {
-                    var component = recipe.Results[i];
-                    var substance = resultSubstances[i] = registry.FromType(component.Type);
-                    resultsToVolume += substance.MillilitersPerMole * component.RelativeAmount;
-                }
-                resultsToVolume /= totalResults;
-
-                // Limit the reaction to prevent spill
-                // TODO: Spill container into surroundings
-                float remainingVolume = container.RemainingVolume;
-                if (maxConversion * totalResults * resultsToVolume - maxConversion * ingredientsToVolume > remainingVolume)
-                {
-                    maxConversion = remainingVolume / resultsToVolume / totalResults;
-                }
-
-                // Remove ingredients
-                for (var i = 0; i < moles.Length; i++)
-                {
-                    container.RemoveSubstance(ingredientSubstances[i], maxConversion * (recipe.Ingredients[i].RelativeAmount / totalIngredients));
-                }
-
-                // Add results
-                foreach (var component in recipe.Results)
-                {
-                    container.AddSubstance(registry.FromType(component.Type), maxConversion * (component.RelativeAmount / totalResults));
-                }
-
-                container.MarkDirty();
-            }
+            substanceContainer.ProcessContainer(container.substanceContainer);
+            container.MarkDirty(); 
         }
     }
 }
